+++ src/jdos/cpu/core_dynamic/EaaBase.java
@@ -1,5 +1,5 @@
 package jdos.cpu.core_dynamic;
 
 public interface EaaBase {
-    public long call();
+    public int call();
 }
